move colors in sub module
