+++ sensorsim/materials/colors.py
@@ -1,0 +1,23 @@
+from enum import Enum
+
+
+class Color(Enum):
+    RED = (1, 0, 0)
+    LIME = (0, 1, 0)
+    BLUE = (0, 0, 1)
+    WHITE = (1, 1, 1)
+    BLACK = (0, 0, 0)
+    YELLOW = (1, 1, 0)
+    CYAN = (0, 1, 1)
+    MAGENTA = (1, 0, 1)
+    SILVER = (0.75, 0.75, 0.75)
+    GRAY = (0.5, 0.5, 0.5)
+    MAROON = (0.5, 0, 0)
+    OLIVE = (0.5, 0.5, 0)
+    GREEN = (0, 0.5, 0)
+    PURPLE = (0.5, 0, 0.5)
+    TEAL = (0, 0.5, 0.5)
+    NAVY = (0, 0, 0.5)
+    CHROME = (0.859, 0.886, 0.914)
+    WOOD = (0.898, 0.729, 0.507)
+    SAFETY_YELLOW = (0.933, 0.824, 0.01)
